Moved serial outside of its own folder
